Update C/C++ Linter Workflow (#59)
This commit updates the C/C++ Linter workflow at `.github/workflows/cpp-linter.yml`. The style is changed from `llvm` to `file` and the action ignores now also `Qt/Qt` and `build/CMakeFiles` folder that are out of the repositories changing reach. As a future restriction, this workflow is only allowed to write status and the remaining is changed to read only.

* Change style to file
* Add Qt to the ignored directories
* Add permissions accordingly for the workflow
* Add `build/CMakeFiles` to the ignore paths